+++ spec/e2e/active_rel_spec.rb
@@ -61,16 +61,21 @@ describe 'ActiveRel' do
       end
     end
 
-    describe '.create!' do
+    describe '#create!' do
       it 'raises an error on invalid params' do
         expect { RelClassWithValidations.create!(from_node: from_node, to_node: to_node) }.to raise_error Neo4j::ActiveRel::Persistence::RelInvalidError
       end
     end
 
-    describe '.save!' do
+    describe '#save!' do
       it 'raises an error on invalid params' do
         invalid_rel = RelClassWithValidations.new(from_node: from_node, to_node: to_node)
         expect { invalid_rel.save! }.to raise_error Neo4j::ActiveRel::Persistence::RelInvalidError
+      end
+
+      it 'returns true on success' do
+        rel = RelClassWithValidations.new(from_node: from_node, to_node: to_node, score: 2)
+        expect(rel.save!).to be true
       end
     end
 
@@ -95,6 +100,51 @@ describe 'ActiveRel' do
         MyRelClass.creates_unique
         MyRelClass.create(from_node: from_node, to_node: to_node)
         expect(from_node.others.count).to eq 1
+      end
+    end
+
+    describe 'type checking' do
+      # rubocop:disable Metrics/AbcSize
+      def self.it_is_expected_to_satisfy(class_method_value)
+        context class_method_value.class.to_s do
+          before { MyRelClass.from_class(class_method_value) }
+
+          it 'fails when given a mismatched value' do
+            expect { MyRelClass.create(from_node: OtherClass.create!, to_node: to_node) }.to raise_error Neo4j::ActiveRel::Persistence::ModelClassInvalidError
+          end
+
+          it 'does not fail when given a matching value' do
+            expect { MyRelClass.create(from_node: FromClass.create, to_node: to_node) }.not_to raise_error
+          end
+        end
+      end
+      # rubocop:enable Metrics/AbcSize
+
+      before { stub_active_node_class('OtherClass') }
+
+      context 'false/:any' do
+        it 'does not check for object/class mismatch' do
+          [false, :any].each do |c|
+            MyRelClass.from_class(c)
+            expect { MyRelClass.create(from_node: OtherClass.create!, to_node: to_node) }.not_to raise_error
+            expect { MyRelClass.create(from_node: from_node, to_node: to_node) }.not_to raise_error
+          end
+        end
+      end
+
+      it_is_expected_to_satisfy('FromClass')
+      it_is_expected_to_satisfy(:FromClass)
+
+      class FromClass; end
+      it_is_expected_to_satisfy(FromClass)
+
+      context 'Array' do
+        # stub_const does not behave with `it_is_expected_to_satisfy` so making it explicit for now...
+        class OtherAcceptableClass
+          include Neo4j::ActiveNode
+        end
+
+        it_is_expected_to_satisfy([OtherAcceptableClass, FromClass])
       end
     end
   end
@@ -197,6 +247,12 @@ describe 'ActiveRel' do
   end
 
   describe 'objects and queries' do
+    around do |ex|
+      ActiveSupport::Deprecation.silenced = true
+      ex.run
+      ActiveSupport::Deprecation.silenced = false
+    end
+
     let!(:rel1) { MyRelClass.create(from_node: from_node, to_node: to_node, score: 99) }
     let!(:rel2) { MyRelClass.create(from_node: from_node, to_node: to_node, score: 49) }